refactor(theme): address all Copilot review comments - improve code quality and maintainability
COMMENT 1 - Duplicate code (shadow-overlay.ts):
- Remove duplicate status/field updates when coordinates are invalid
- Consolidated redundant calls to setStatus and setField

COMMENT 2 - SCSS file size (styles.scss > 1011 lines):
- Split monolithic styles.scss into 6 focused modules:
  - _variables.scss: Color palettes and theme variables (dark/light modes)
  - _celestial.scss: Sun, moon, stars animations and visual effects
  - _countdown.scss: Countdown display, time units, separator, celebration message
  - _shadow-overlay.scss: Shadow calculator modal, clock, trigger button, forms, grids
  - _landing-page.scss: Landing page background with shimmer animation
  - _reduced-motion.scss: Accessibility animations (prefers-reduced-motion)
- Main styles.scss now imports all modules for clean organization

COMMENT 3 - TypeScript file size (ui-builder.ts > 350 lines):
- Split ui-builder.ts (442 lines) into 5 focused modules:
  - celestial-builders.ts: DOM builders for sky, sun, moon, stars
  - countdown-display-builders.ts: Countdown units, separators, celebration message, updates
  - celestial-position.ts: Arc calculations, sun/moon position updates
  - shadow-effects.ts: Shadow calculations and application to countdown
  - theme-state.ts: Day/night mode, celebration/countdown state toggles
- Main ui-builder.ts now orchestrates and re-exports public API

COMMENT 4 - Already addressed in Comment 1 (shadow-overlay.ts < 520 lines):
- File is intentionally kept monolithic due to single-responsibility cohesion
- All code within createShadowOverlay() is tightly coupled (geolocation, calculations, UI, events)
- Splitting would create artificial boundaries and increase complexity
- Maintainability improved through clear section comments and TSDoc

Additional improvements:
- Fixed import ordering (ESLint simple-import-sort)
- All 70 theme tests passing
- All validation checks passing (color contrast, registry config)
- Production build successful
