+++ src/themes/akhil-square-shadow/_countdown.scss
@@ -1,0 +1,145 @@
+/**
+ * Countdown Display
+ *
+ * Styles for countdown digits, separators, and celebration message.
+ */
+
+.akhil-square-shadow-theme {
+  position: fixed;
+  inset: 0;
+  width: 100vw;
+  height: 100vh;
+  overflow: hidden;
+
+  // Center countdown content
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+
+  // Respect safe area
+  padding: var(--safe-area-top, 60px) 20px var(--safe-area-bottom, 20px);
+  box-sizing: border-box;
+
+  color: var(--shadow-theme-text);
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Helvetica, Arial, sans-serif;
+
+  // Day mode is default
+  &:not(.is-night) {
+    .shadow-theme-sky {
+      background: var(--shadow-theme-day-sky);
+    }
+  }
+
+  // Night mode
+  &.is-night {
+    .shadow-theme-sky {
+      background: var(--shadow-theme-night-sky);
+    }
+
+    .shadow-countdown-display {
+      color: #ffffff;
+    }
+  }
+
+  // Celebration mode
+  &.is-celebrating {
+    .shadow-theme-sky {
+      background: linear-gradient(
+        180deg,
+        #1a0a2e 0%,
+        #2d1a4a 20%,
+        #4a2a6a 40%,
+        #6a3a8a 60%,
+        #8a4aaa 80%,
+        #aa5aca 100%
+      );
+    }
+
+    .shadow-countdown-display,
+    .shadow-celebration-message {
+      color: var(--shadow-theme-celebration);
+      text-shadow:
+        0 0 20px var(--shadow-theme-celebration),
+        0 0 40px var(--shadow-theme-celebration),
+        0 0 60px var(--shadow-theme-celebration);
+    }
+  }
+}
+
+// =============================================================================
+// Countdown Display
+// =============================================================================
+
+.shadow-countdown-display {
+  display: flex;
+  align-items: center;
+  gap: clamp(12px, 3vw, 24px);
+  font-size: clamp(3rem, calc(6rem * var(--font-scale, 1)), 8rem);
+  font-weight: 700;
+  text-align: center;
+  z-index: 20;
+  color: var(--shadow-theme-text);
+
+  // Base text shadow (will be overridden by JS for dynamic shadow)
+  text-shadow: 2px 4px 8px var(--shadow-theme-text-shadow);
+
+  // Smooth shadow transition
+  transition: text-shadow 0.1s ease-out;
+}
+
+.shadow-time-unit {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 0.5rem;
+
+  .shadow-value {
+    font-variant-numeric: tabular-nums;
+    line-height: 1;
+  }
+}
+
+.shadow-separator {
+  opacity: 0.6;
+  font-weight: 400;
+  line-height: 1;
+  animation: blink 1s step-end infinite;
+}
+
+@keyframes blink {
+  0%,
+  100% {
+    opacity: 0.6;
+  }
+  50% {
+    opacity: 0.2;
+  }
+}
+
+// =============================================================================
+// Celebration Message
+// =============================================================================
+
+.shadow-celebration-message {
+  font-size: clamp(2.5rem, calc(5rem * var(--font-scale, 1)), 6rem);
+  font-weight: 700;
+  text-align: center;
+  z-index: 20;
+  margin: 0;
+  padding: 1rem 2rem;
+  color: var(--shadow-theme-celebration);
+
+  animation: celebration-appear 0.8s cubic-bezier(0.34, 1.56, 0.64, 1) forwards;
+}
+
+@keyframes celebration-appear {
+  from {
+    opacity: 0;
+    transform: scale(0.5) translateY(20px);
+  }
+  to {
+    opacity: 1;
+    transform: scale(1) translateY(0);
+  }
+}
